requirements.txt
Flask==3.1.0
pandas==2.3.1
fpdf2
openpyxl